api/requirements.txt
pandas
requests
flask
flask_cors
setuptools
kiteconnect
pyotp
undetected_chromedriver
selenium
fyers-apiv3
botocore==1.34.7
psycopg2-binary
ta
openpyxl
xlsxwriter
jwt